// src/app/core/models/cda/applicationFileMaster.ts
export class ApplicationFileMaster {
    applicationFileMasterID: number;
    refNo: string;
    applicantName: string;
    approvalDate: string;
    applicationType: ApplicationType;
    applicationStatus: ApplicationStatus;
    rsNo: string;
    bsNo: string;
    thanaID: number;
    mouzaID: number;
    dpzid: number;
    road: string;
    officerInChargeID: number | null;
    isVisited: boolean;
    latitude: number | null;
    longitude: number | null;
    createdBy: number;
    createdDate: string;
    updatedBy: number;
    updatedByFullName: string;
    updatedByFullNameBangla: string;
    updatedDate: string;
    approvalDateSt: string;
    thanaName: string;
    thanaNameBangla: string;
    mouzaName: string;
    mouzaNameBangla: string;
    dpzName: string;
    dpzNameBangla: string;
    userFullName: string;
    assignBy: number;
    assignByName: string;
    assignByNameBangla: string;
    targetDate: string;
    assignDate: string;
    applicationTypeName: string;
    applicationStatusName: string
    investigationOfficerID: number;
    investigationOfficerName: string;
    tR_InvestigationOfficerID: number;
    tR_InvestigationOfficerName: string;
    inspectionDetailsBCSPCase: InspectionDetailsBCSPCase;
    inspectionDetailsTechnical: InspectionDetailsBCSPCase;
    luc: LUC;
    visitDate?: Date;
    invOfficerFullNameBangla: string;

    isChecked: boolean;
    isFileDetails: boolean = false;
    applicationTypeID: number;
    applicationTypeSt: string;
    case_id:string;
    case_no:string;
    applicant_father:string;
    dap_zoing:string;
    occupancy_use:string;
    occupancy_type:string;
    road_width_north_foot:string;
    road_width_south_foot:string;
    road_width_east_foot:string;
    road_width_west_foot:string;
    around_condition:string;
    road_sugg_north_foot:string;
    road_sugg_south_foot:string;
    road_sugg_east_foot:string;
    road_sugg_west_foot:string;
    dap_desc:string;
    land_uses:string;
    other_desc:string;
    inspector_name:string;
    applicant_mobile:string;
    land_owner_name:string;
    land_lease:string;
    construction_measure_south:string;
    construction_measure_north:string;
    construction_measure_east:string;
    construction_measure_west:string;
    land_quantity:string;
    is_applicant_land_owner:string;
    tafsil:string;
    rs_ledger_number:string;
    thana:string;
    mouza_name:string;
    layout_plot_no:string;
    layout_area:string;
    first_floor:string;
    second_floor_from:string;
    second_floor_to:string;
    shiri:string;
    total_floor:string;
    parking_quantity:string;
    nature:string;
    demolished_quantity:string;
    open_space_exist:string;
    road_name:string;
    road_wide:string;
    road_direction:string;
    road_recognized:string;
    road_traffic_recognized:string;
    is_work_start:string;
    previous_approval_details:string;
    submitted_notice_details:string;
    land_description:string;
    main_road_name:string;
    key_point_installation:string;
    is_mountain_land:string;
    distance_mountain:string;
    is_land_agriculture:string;
    structure_within_12meter:string;
    structure_north:string;
    structure_south:string;
    structure_east:string;
    structure_west:string;
    beside_canal:string;
    canal_wide:string;
    canal_ripe_or_raw:string;
    water_drainage:string;
    drainage_connection:string;
    electric_structure:string;
    electric_line_length:string;
    defense_section_structure:string;
    proposed_area_pond:string;
    proposed_area_historical:string;
    site_created_by_pond_fill:string;
    construction_side_effect:string;
    luc_no:string;
    spcase_no:string;

    landOwnerName: string;
    landLease: string;
    ConstructionMeasureSouth: string;
    constructionMeasureNorth: string;
    constructionMeasureEast: string;
    constructionMeasureWest: string;
    landQuantity: string;
    isApplicantLandOwner: string;
    rSLedgerNumber: string;
    layoutPlotNo: string;
    layoutArea: string;
    mainRoadName: string;
    totalFloor: string;
    roadName: string;
}