// src/app/core/models/settings/userDpzMap.ts
export class UserDPZMap {
    userDPZMapID: number;
    userID: number = undefined;
    dpzid: number;
    departmentID: number;
    isActive: boolean;
    fromDate: string;
    toDate: string;
    createdBy: number;
    createdDate: string;
    updatedBy: number;
    updatedDate: string;
    userName: string;
    userNameBangla: string;
    dpzName: string;
    dpzNameBangla: string;
    departmentName: string;
}